Create Update Windows machines and send notifications to Logz.io
This script is written in PowerShell and is intended to monitor a Windows machine for certain events and alert the user if any of these events occur. The script first imports the AWS PowerShell module and sets the AWS credentials and default region. It then retrieves the secret value from AWS Secrets Manager and converts it to a secure string. Next, the script sets the URL for the logz.io listener and retrieves any available updates for the machine. If there are updates available, the script installs them and restarts the machine.